+++ packages/tinkoff_acquiring/lib/src/core/models/add_card/add_card_request.dart
@@ -1,6 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../constants.dart';
 import '../base/acquiring_request.dart';
 import '../enums/check_type.dart';
