+++ packages/tinkoff_acquiring/lib/src/core/models/attach_card/attach_card_request.dart
@@ -1,6 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../constants.dart';
 import '../base/acquiring_request.dart';
 
 part 'attach_card_request.g.dart';
