+++ packages/tinkoff_acquiring/lib/src/core/models/cancel/cancel_request.dart
@@ -1,7 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../constants.dart';
-import '../../../utils/extensions.dart';
 import '../base/acquiring_request.dart';
 import '../common/receipt.dart';
