+++ packages/tinkoff_acquiring/lib/src/core/models/check_3ds_version/check_3ds_version_request.dart
@@ -1,7 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../constants.dart';
-import '../../../utils/extensions.dart';
 import '../base/acquiring_request.dart';
 
 part 'check_3ds_version_request.g.dart';
