+++ packages/tinkoff_acquiring/lib/src/core/models/get_card_list/get_card_list_request.dart
@@ -1,6 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../constants.dart';
 import '../base/acquiring_request.dart';
 
 part 'get_card_list_request.g.dart';
