+++ packages/tinkoff_acquiring/lib/src/core/models/get_static_qr/get_static_qr_request.dart
@@ -1,6 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../constants.dart';
 import '../base/acquiring_request.dart';
 import '../enums/data_type.dart';
