+++ packages/tinkoff_acquiring/lib/src/core/models/remove_customer/remove_customer_request.dart
@@ -1,6 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../constants.dart';
 import '../base/acquiring_request.dart';
 
 part 'remove_customer_request.g.dart';
