+++ packages/tinkoff_acquiring/lib/src/core/models/resend/resend_request.dart
@@ -1,6 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../constants.dart';
 import '../base/acquiring_request.dart';
 
 part 'resend_request.g.dart';
