+++ packages/tinkoff_acquiring/lib/src/core/network/network_client.dart
@@ -4,7 +4,6 @@ import 'dart:convert';
 import 'package:http/http.dart' as http;
 import 'package:meta/meta.dart';
 
-import '../../constants.dart';
 import '../../utils/crypto_utils.dart';
 import '../models/base/acquiring_request.dart';
 import '../models/base/acquiring_response.dart';
